Move inspectableT helper to utility independent file
